+++ dbs.txt
@@ -1,1 +1,3 @@
 aaa
+bbb
+all
